move `GET /cases` endpoint to supervision controller and update test cases
